snakep2p: handling of a stub event channel.
